+++ src/main/scala/com/es/ml/recommendation/ALSTrain.scala
@@ -70,11 +70,11 @@ object ALSTrain {
 
   }
 
-  case class Rating(userId: Int, movieId: Int, rating: Float, timestamp: Long)
+  case class Rating(userId: Int, movieId: Int, rating: Double, timestamp: Long)
   def parseRating(str: String): Rating = {
     val fields = str.split("::")
     assert(fields.size == 4)
-    Rating(fields(0).toInt, fields(1).toInt, fields(2).toFloat, fields(3).toLong)
+    Rating(fields(0).toInt, fields(1).toInt, fields(2).toDouble, fields(3).toLong)
   }
 
   def run(p:Params): Unit = {
